+++ bbp/init.lua
@@ -1,0 +1,6 @@
+local bbp = {}
+
+bbp.utils = require("bbp.utils")
+bbp.modloader = require("bbp.modloader")
+
+return bbp
