+++ bbp/mod-loader.lua
@@ -1,52 +1,4 @@
-bbp = {
-	mods = {}
-}
--- TODO: remove this later due to deprecation
-mods = bbp.mods
-
-local function readJsonFromFile(filePath)
-	local contents = love.filesystem.read(filePath)
-	return json.decode(contents)
-end
-
--- files.walk implementation
-local function loopFiles(tbl, dir, callback)
-	local function recurse(currentTable, currentDir)
-		for _, item in ipairs(love.filesystem.getDirectoryItems(currentDir)) do
-			local fullPath = currentDir .. "/" .. item
-			if love.filesystem.getInfo(fullPath, "directory") then
-				currentTable[item] = currentTable[item] or {}
-				recurse(currentTable[item], fullPath)
-			else
-				local fileName = item:match("([^/]+)$"):gsub("%..+$", "") -- extract the file name without the directory and extension
-				callback(currentTable, fullPath, fileName)
-			end
-		end
-	end
-
-	recurse(tbl, dir)
-end
-
-local function moveDirectory(source, target)
-    love.filesystem.createDirectory(target)
-
-    local files = love.filesystem.getDirectoryItems(source)
-    for _, file in pairs(files) do
-        local sourceFilePath = source .. file
-        local targetFilePath = target .. file
-
-        local contents = love.filesystem.read(sourceFilePath)
-        if contents then
-            local targetFile = love.filesystem.newFile(targetFilePath)
-            targetFile:open("w")
-            targetFile:write(contents)
-            targetFile:flush()
-            love.filesystem.remove(sourceFilePath)
-        end
-    end
-
-    love.filesystem.remove(source)
-end
+local modloader = {}
 
 local function setModChunkEnvironment(chunk, mod, setDeprecated)
 	local env = setmetatable({}, {
@@ -80,23 +32,6 @@ local function setModChunkEnvironment(chunk, mod, setDeprecated)
 	return setfenv(chunk, env)
 end
 
-local function printTable(t, indent, title)
-	if title then
-		print(title)
-	end
-	indent = indent or 0
-	local indentStr = string.rep("  ", indent)
-
-	for k, v in pairs(t) do
-		if type(v) == "table" then
-			print(indentStr .. tostring(k) .. ":")
-			printTable(v, indent + 1)
-		else
-			print(indentStr .. tostring(k) .. ": " .. tostring(v))
-		end
-	end
-end
-
 local function mergeLangFiles(originalLoc, modLoc)
 	local selectedLanguage = savedata.options.language
 	for key, value in pairs(modLoc) do
@@ -107,22 +42,6 @@ local function mergeLangFiles(originalLoc, modLoc)
 	end
 end
 
-function tableContains(t, v)
-	for i = 1, #t do
-		if (t[i] == v) then
-			return true
-		end
-	end
-	return false
-end
-
-local function getParent(fullPath)
-	return fullPath:match("(.*/)") or "."
-end
-
-function string:endswith(ending)
-	return ending == "" or self:sub(- #ending) == ending
-end
 
 local function getModConfigRenderer(mod)
 	if mod._configRenderer == false then
@@ -156,22 +75,32 @@ local function setModEnabled(mod, enabled)
 	if enabled == nil then
 		enabled = true
 	end
+
 	local createFilePath = mod.path .. (enabled and "/.nolovelyignore" or "/.lovelyignore")
 	local deleteFilePath = mod.path .. (enabled and "/.lovelyignore" or "/.nolovelyignore")
 	local success = true
+
 	if not love.filesystem.getInfo(createFilePath, 'file') then
 		success = success and love.filesystem.write(createFilePath, "")
 	end
+
 	if love.filesystem.getInfo(deleteFilePath, 'file') then
 		success = success and love.filesystem.remove(deleteFilePath)
 	end
+
 	if not success then
 		return
 	end
+
 	rawset(mod, '_enabled', enabled)
 end
 
-function loadMods() -- loads mod data, assets, mod icons etc.
+function modloader.loadMods() -- loads mod data, assets, mod icons etc.
+	bbp.mods = {}
+
+	-- TODO: remove this later due to deprecation
+	mods = bbp.mods
+
 	local modsPath = "Mods"
 	local success = love.filesystem.getInfo(modsPath, 'directory')
 
@@ -226,7 +155,7 @@ function loadMods() -- loads mod data, assets, mod icons etc.
 
 		-- load mod data if it exists
 		if love.filesystem.getInfo(mod.path .. "/mod.json", 'file') then
-			local modData = readJsonFromFile(mod.path .. "/mod.json")
+			local modData = dpf.loadJson(mod.path .. "/mod.json")
 			mod.id = modData.id or mod.id
 			mod.name = modData.name or mod.name
 			mod.author = modData.author or mod.author
@@ -286,25 +215,25 @@ function loadMods() -- loads mod data, assets, mod icons etc.
 		local assetsPath = mod.path .. "/assets"
 		if love.filesystem.getInfo(assetsPath, 'directory') then
 			-- load sprites
-			loopFiles(sprites, assetsPath .. "/textures", function(tbl, path, fileName)
+			bbp.utils.loopFiles(sprites, assetsPath .. "/textures", function(tbl, path, fileName)
 				print("[BB+] injecting sprite " .. path .. "...")
 				tbl[fileName] = love.graphics.newImage(path)
 			end)
 
 			-- load sounds
-			loopFiles(sounds, assetsPath .. "/sounds", function(tbl, path, fileName)
+			bbp.utils.loopFiles(sounds, assetsPath .. "/sounds", function(tbl, path, fileName)
 				print("[BB+] injecting sound " .. path .. "...")
 				tbl[fileName] = love.sound.newSoundData(path)
 			end)
 
 			-- load shaders
-			loopFiles(shaders, assetsPath .. "/shaders", function(tbl, path, fileName)
+			bbp.utils.loopFiles(shaders, assetsPath .. "/shaders", function(tbl, path, fileName)
 				print("[BB+] injecting shader " .. path .. "...")
 				tbl[fileName] = love.graphics.newShader(path)
 			end)
 
 			-- load animations
-			loopFiles(animations, assetsPath .. "/animations", function(tbl, path, fileName)
+			bbp.utils.loopFiles(animations, assetsPath .. "/animations", function(tbl, path, fileName)
 				if path:endswith(".png") then
 					print("[BB+] injecting animation " .. path .. "...")
 					local data = getParent(path) .. "data.json"
@@ -315,7 +244,8 @@ function loadMods() -- loads mod data, assets, mod icons etc.
 				end
 			end)
 
-			loopFiles(loc.json, assetsPath .. "/lang", function(tbl, path, fileName)
+			-- load lang files
+			bbp.utils.loopFiles(loc.json, assetsPath .. "/lang", function(tbl, path, fileName)
 				table.insert(customLanguages, fileName)
 				-- make sure we don't load english lang when owo is selected
 				if fileName == savedata.options.language then
@@ -327,7 +257,7 @@ function loadMods() -- loads mod data, assets, mod icons etc.
 		end
 
 		-- load states
-		loopFiles({}, mod.path .. "/states", function(_, path, fileName)
+		bbp.utils.loopFiles({}, mod.path .. "/states", function(_, path, fileName)
 			print("[BB+] injecting state " .. path .. "...")
 			bs.fromPath(fileName, path)
 			if bs.states[fileName] then
@@ -336,7 +266,7 @@ function loadMods() -- loads mod data, assets, mod icons etc.
 		end)
 
 		-- load entities
-		loopFiles({}, mod.path .. "/entities", function(_, path, fileName)
+		bbp.utils.loopFiles({}, mod.path .. "/entities", function(_, path, fileName)
 			print("[BB+] injecting entity " .. path .. "...")
 			em.new(path, fileName)
 			if em.entities[fileName] then
@@ -357,18 +287,10 @@ function loadMods() -- loads mod data, assets, mod icons etc.
 	end
 
 	print("[BB+] Finished loading all mods! :D")
-	printTable(animations, 3, "Animations:")
-	printTable(sprites, 3, "Sprites:")
-	printTable(sounds, 3, "Sounds:")
-	printTable(shaders, 3, "Shaders:")
+	bbp.utils.printTable(animations, "Animations:")
+	bbp.utils.printTable(sprites, "Sprites:")
+	bbp.utils.printTable(sounds, "Sounds:")
+	bbp.utils.printTable(shaders, "Shaders:")
 end
 
-function getModNames()
-	local modNames = {}
-
-	for _, mod in pairs(bbp.mods) do
-		table.insert(modNames, "  - " .. mod.name .. " (" .. mod.version .. ") by " .. mod.author)
-	end
-
-	return table.concat(modNames, "\n")
-end
+return modloader
